General changes, the game has a main menu now

+++ build.ts
@@ -1,12 +1,17 @@
-import { $ } from "bun"
+import { $ } from "bun";
+import fs from "node:fs";
+(async () => {
+	await $`bun tailwindcss -i ./index.css -o ./dist.css`;
 
-await $`bun tailwindcss -i ./index.css -o ./dist.css`
-await $`rm -r dist/*`
-await Bun.build({
-    entrypoints: ["./index.html"],
-    outdir: "./dist",
-    minify: true,
-    define: {
-        ["process.env.NODE_ENV"]: '"production"',
-    }
-})
+	await $`mkdir -p dist`;
+	if (fs.existsSync("dist") && fs.existsSync("dist/index.html"))
+		await $`rm -r dist/*`;
+	await Bun.build({
+		entrypoints: ["./index.html"],
+		outdir: "./dist",
+		minify: true,
+		define: {
+			["process.env.NODE_ENV"]: '"production"',
+		},
+	});
+})();
